IrrigationControl/requirements.txt
requests
paho-mqtt
influxdb
apscheduler
cherrypy